Refactor DoT resolver
